Moved code from ImportExportTest into ExportTests
